+++ libMx/d_auto_amplifier.c
@@ -540,7 +540,7 @@ mxd_auto_amplifier_change_control( MX_AUTOSCALE *autoscale )
 	old_gain = auto_amplifier->present_amplifier_gain;
 
 	MX_DEBUG( 2,("%s: change_control = %d",
-			fname, autoscale->change_control));
+			fname, (int) autoscale->change_control));
 
 	switch( autoscale->change_control ) {
 	case MXF_AUTO_NO_CHANGE:
@@ -575,7 +575,7 @@ mxd_auto_amplifier_change_control( MX_AUTOSCALE *autoscale )
 	default:
 		return mx_error( MXE_ILLEGAL_ARGUMENT, fname,
 			"Illegal value %d for change control.",
-			autoscale->change_control );
+			(int) autoscale->change_control );
 	}
 
 	MX_DEBUG( 2,("%s: Changing gain from %g to %g",
@@ -690,7 +690,7 @@ mxd_auto_amplifier_get_parameter( MX_AUTOSCALE *autoscale )
 		fname, autoscale->record->name,
 		mx_get_field_label_string( autoscale->record,
 			autoscale->parameter_type ),
-		autoscale->parameter_type));
+		(int) autoscale->parameter_type));
 
 	switch( autoscale->parameter_type ) {
 	default:
@@ -720,7 +720,7 @@ mxd_auto_amplifier_set_parameter( MX_AUTOSCALE *autoscale )
 		fname, autoscale->record->name,
 		mx_get_field_label_string( autoscale->record,
 			autoscale->parameter_type ),
-		autoscale->parameter_type));
+		(int) autoscale->parameter_type));
 
 	switch( autoscale->parameter_type ) {
 	default:
