+++ libMx/d_cm17a_doutput.c
@@ -255,7 +255,7 @@ mxd_cm17a_doutput_open( MX_RECORD *record )
 		return mx_error( MXE_ILLEGAL_ARGUMENT, fname,
 	"Illegal device code %d selected for CM17A digital output '%s'.  "
 	"The allowed values are from 1 to 16.",
-			cm17a_doutput->device_code, record->name );
+			(int) cm17a_doutput->device_code, record->name );
 	}
 
 	/* Compute the values of all the commands. */
