+++ libMx/d_compumotor_linear.c
@@ -8,7 +8,7 @@
  *
  *------------------------------------------------------------------------
  *
- * Copyright 1999-2004 Illinois Institute of Technology
+ * Copyright 1999-2004, 2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -853,7 +853,7 @@ mxd_compumotor_linear_get_position( MX_MOTOR *motor )
 	n = compumotor_linear_motor->controller_index;
 
 	sprintf( command, "%d_!TPE",
-			compumotor_interface->controller_number[n] );
+			(int) compumotor_interface->controller_number[n] );
 
 	mx_status = mxi_compumotor_command( compumotor_interface, command,
 			response, sizeof(response), COMPUMOTOR_LINEAR_DEBUG );
@@ -964,7 +964,8 @@ mxd_compumotor_linear_soft_abort( MX_MOTOR *motor )
 
 	n = compumotor_linear_motor->controller_index;
 
-	sprintf( command, "%d_!S", compumotor_interface->controller_number[n] );
+	sprintf( command, "%d_!S",
+			(int) compumotor_interface->controller_number[n] );
 
 	for ( i = 0; i < compumotor_linear_motor->num_axes; i++ ) {
 
@@ -1025,7 +1026,8 @@ mxd_compumotor_linear_immediate_abort( MX_MOTOR *motor )
 
 	n = compumotor_linear_motor->controller_index;
 
-	sprintf(command, "%d_!K", compumotor_interface->controller_number[n]);
+	sprintf(command, "%d_!K",
+		(int) compumotor_interface->controller_number[n]);
 
 	mx_status = mxi_compumotor_command( compumotor_interface, command,
 					NULL, 0, COMPUMOTOR_LINEAR_DEBUG );
